+++ packages/widget/src/components/Messages/useMessageQueue.ts
@@ -2,6 +2,7 @@ import type { Route } from '@lifi/sdk'
 import { useMemo } from 'react'
 import { useFromTokenSufficiency } from '../../hooks/useFromTokenSufficiency.js'
 import { useGasSufficiency } from '../../hooks/useGasSufficiency.js'
+import { useRouteRequiredAccountConnection } from '../../hooks/useRouteRequiredAccountConnection.js'
 import { useToAddressRequirements } from '../../hooks/useToAddressRequirements.js'
 
 interface QueuedMessage {
@@ -15,27 +16,38 @@ export const useMessageQueue = (route?: Route, allowInteraction?: boolean) => {
     requiredToAddress,
     toAddress,
     accountNotDeployedAtDestination,
+    accountDeployedAtDestination,
     isLoading: isToAddressRequirementsLoading,
   } = useToAddressRequirements(route)
   const { insufficientFromToken, isLoading: isFromTokenSufficiencyLoading } =
     useFromTokenSufficiency(route)
   const { insufficientGas, isLoading: isGasSufficiencyLoading } =
     useGasSufficiency(route)
+  const { missingChain, missingAccountAddress } =
+    useRouteRequiredAccountConnection(route)
 
   const messageQueue = useMemo(() => {
     const queue: QueuedMessage[] = []
 
+    if (missingChain) {
+      queue.push({
+        id: 'MISSING_ROUTE_REQUIRED_ACCOUNT',
+        priority: 1,
+        props: { chain: missingChain, address: missingAccountAddress },
+      })
+    }
+
     if (insufficientFromToken) {
       queue.push({
         id: 'INSUFFICIENT_FUNDS',
-        priority: 1,
+        priority: 2,
       })
     }
 
     if (insufficientGas?.length) {
       queue.push({
         id: 'INSUFFICIENT_GAS',
-        priority: 2,
+        priority: 3,
         props: { insufficientGas },
       })
     }
@@ -43,14 +55,21 @@ export const useMessageQueue = (route?: Route, allowInteraction?: boolean) => {
     if (accountNotDeployedAtDestination && !allowInteraction) {
       queue.push({
         id: 'ACCOUNT_NOT_DEPLOYED',
-        priority: 3,
+        priority: 4,
       })
     }
 
     if (requiredToAddress && !toAddress) {
       queue.push({
         id: 'TO_ADDRESS_REQUIRED',
-        priority: 4,
+        priority: 5,
+      })
+    }
+
+    if (accountDeployedAtDestination && !allowInteraction) {
+      queue.push({
+        id: 'ACCOUNT_DEPLOYED',
+        priority: 6,
       })
     }
 
@@ -59,9 +78,12 @@ export const useMessageQueue = (route?: Route, allowInteraction?: boolean) => {
     allowInteraction,
     insufficientFromToken,
     insufficientGas,
+    accountDeployedAtDestination,
     accountNotDeployedAtDestination,
     requiredToAddress,
     toAddress,
+    missingChain,
+    missingAccountAddress,
   ])
 
   return {