+++ packages/widget/src/hooks/useToAddressRequirements.ts
@@ -3,12 +3,12 @@ import { useAccount } from '@lifi/wallet-management'
 import { useChain } from '../hooks/useChain.js'
 import { useWidgetConfig } from '../providers/WidgetProvider/WidgetProvider.js'
 import { useFieldValues } from '../stores/form/useFieldValues.js'
-import { RequiredUI } from '../types/widget.js'
+import { HiddenUI, RequiredUI } from '../types/widget.js'
 import { isDelegationDesignatorCode } from '../utils/eip7702.js'
 import { useIsContractAddress } from './useIsContractAddress.js'
 
 export const useToAddressRequirements = (route?: RouteExtended) => {
-  const { requiredUI } = useWidgetConfig()
+  const { requiredUI, hiddenUI } = useWidgetConfig()
   const [formFromChainId, formToChainId, formToAddress] = useFieldValues(
     'fromChain',
     'toChain',
@@ -57,9 +57,10 @@ export const useToAddressRequirements = (route?: RouteExtended) => {
     !fromContractCodeHasDelegationIndicator
 
   const requiredToAddress =
-    requiredUI?.includes(RequiredUI.ToAddress) ||
-    isDifferentChainType ||
-    isCrossChainContractAddress
+    (isDifferentChainType ||
+      isCrossChainContractAddress ||
+      requiredUI?.includes(RequiredUI.ToAddress)) &&
+    !hiddenUI?.includes(HiddenUI.ToAddress)
 
   const accountNotDeployedAtDestination =
     isFromContractAddress &&
@@ -67,10 +68,17 @@ export const useToAddressRequirements = (route?: RouteExtended) => {
     !isToContractAddress &&
     fromAddress?.toLowerCase() === toAddress?.toLowerCase()
 
+  const accountDeployedAtDestination =
+    isFromContractAddress &&
+    isToContractAddress &&
+    !fromContractCodeHasDelegationIndicator &&
+    fromAddress?.toLowerCase() === toAddress?.toLowerCase()
+
   return {
     requiredToAddress,
     requiredToChainType: toChain?.chainType,
     accountNotDeployedAtDestination,
+    accountDeployedAtDestination,
     toAddress,
     isFromContractAddress,
     isToContractAddress,