+++ packages/widget/src/hooks/useTokenBalances.ts
@@ -2,8 +2,10 @@ import { getTokenBalances } from '@lifi/sdk'
 import { useAccount } from '@lifi/wallet-management'
 import { useQuery } from '@tanstack/react-query'
 import { formatUnits } from 'viem'
+import { useWidgetConfig } from '../providers/WidgetProvider/WidgetProvider.js'
 import type { FormType } from '../stores/form/types.js'
 import type { TokenAmount } from '../types/token.js'
+import { getQueryKey } from '../utils/queries.js'
 import { useTokens } from './useTokens.js'
 
 const defaultRefetchInterval = 32_000
@@ -17,6 +19,7 @@ export const useTokenBalances = (
     formType
   )
   const { account } = useAccount({ chainType: chain?.chainType })
+  const { keyPrefix } = useWidgetConfig()
 
   const isBalanceLoadingEnabled =
     Boolean(account.address) &&
@@ -29,7 +32,7 @@ export const useTokenBalances = (
     refetch,
   } = useQuery({
     queryKey: [
-      'token-balances',
+      getQueryKey('token-balances', keyPrefix),
       account.address,
       selectedChainId,
       tokens?.length,