+++ packages/widget/src/hooks/useTokenSearch.ts
@@ -3,6 +3,7 @@ import { useQuery, useQueryClient } from '@tanstack/react-query'
 import { useWidgetConfig } from '../providers/WidgetProvider/WidgetProvider.js'
 import type { FormType } from '../stores/form/types.js'
 import type { TokenAmount } from '../types/token.js'
+import { getQueryKey } from '../utils/queries.js'
 
 export const useTokenSearch = (
   chainId?: number,
@@ -11,10 +12,10 @@ export const useTokenSearch = (
   formType?: FormType
 ) => {
   const queryClient = useQueryClient()
-  const { tokens: configTokens } = useWidgetConfig()
+  const { tokens: configTokens, keyPrefix } = useWidgetConfig()
 
   const { data, isLoading } = useQuery({
-    queryKey: ['token-search', chainId, tokenQuery],
+    queryKey: [getQueryKey('token-search', keyPrefix), chainId, tokenQuery],
     queryFn: async ({ queryKey: [, chainId, tokenQuery], signal }) => {
       const token = await getToken(chainId as ChainId, tokenQuery as string, {
         signal,
@@ -37,7 +38,7 @@ export const useTokenSearch = (
         }
 
         queryClient.setQueriesData<TokensResponse>(
-          { queryKey: ['tokens'] },
+          { queryKey: [getQueryKey('tokens', keyPrefix)] },
           (data) => {
             if (
               data &&